Backend/requirements.txt
blinker==1.9.0
click==8.1.8
celery==5.5.1
eventlet
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2
Jinja2==3.1.5
joblib
MarkupSafe==3.0.2
mediapipe==0.10.21
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.1
SQLAlchemy==2.0.38
Werkzeug==3.1.3
pymongo==4.10.1
flask-pymongo==2.3.0
flask-jwt-extended==4.6.0
bcrypt==4.2.1
python-dotenv==1.0.1
flask-cors==5.0.0